+++ libcxx/test/std/language.support/support.dynamic/new.delete/new.delete.array/new.size.except.pass.cpp
@@ -7,6 +7,8 @@
 //===----------------------------------------------------------------------===//
 
 // UNSUPPORTED: no-exceptions
+// UNSUPPORTED: sanitizer-new-delete
+
 #include <new>
 #include <cassert>
 #include <limits>
@@ -29,9 +31,9 @@ void my_new_handler() {
 int main(int, char**) {
   std::set_new_handler(my_new_handler);
   try {
-    void* x = operator new[](std::numeric_limits<std::size_t>::max(), std::nothrow);
+    void* x = operator new[](std::numeric_limits<std::size_t>::max());
     (void)x;
-    assert(x == NULL);
+    assert(false);
   } catch (my_bad_alloc const& e) {
     assert(new_handler_called == 1);
     assert(e.self == &e);
