+++ libcxx/test/std/language.support/support.dynamic/new.delete/new.delete.single/new.size_align_nothrow.except.pass.cpp
@@ -7,6 +7,8 @@
 //===----------------------------------------------------------------------===//
 
 // UNSUPPORTED: no-exceptions
+// UNSUPPORTED: sanitizer-new-delete
+
 #include <new>
 #include <cassert>
 #include <limits>
@@ -29,7 +31,7 @@ void my_new_handler() {
 int main(int, char**) {
   std::set_new_handler(my_new_handler);
   try {
-    void* x = operator new(std::numeric_limits<std::size_t>::max(), std::nothrow);
+    void* x = operator new(std::numeric_limits<std::size_t>::max(), static_cast<std::align_val_t>(32), std::nothrow);
     (void)x;
     assert(x == NULL);
   } catch (my_bad_alloc const& e) {
